refactor(codegen): extract std::io::fs::* into stdlib/io_fs.rs (Round 1.10)
20 fns lifted into IoFsStdlib<'ctx>: 10 fallible (read_file, read_bytes,
write_file, write_file_append, file_size, mkdir, rename, unlink,
mktemp, list_dir_count, list_dir_at) and 10 non-fallible (the
historical sister fns retained from pre-IoError shape, plus the
predicates file_exists / extension).

Surface bumps to pub(crate):
- Cx method: current_arena_ptr

codegen.rs: 43434 -> 42235 (-1199 lines). Largest single sub-round so far.
