+++ crates/hale-codegen/src/stdlib/mod.rs
@@ -12,6 +12,7 @@
 pub(crate) mod crypto;
 pub(crate) mod decimal;
 pub(crate) mod env;
+pub(crate) mod io_fs;
 pub(crate) mod io_stdin;
 pub(crate) mod math;
 pub(crate) mod process;
